sxiv only if not installed

+++ cookbooks/sxiv/recipes/default.rb
@@ -6,6 +6,7 @@ git "#{Chef::Config[:file_cache_path]}/sxiv" do
 	repository "https://github.com/muennich/sxiv.git"
 	action :checkout
 	notifies :run, "execute[install sxiv]", :immediately
+	not_if { ::File.exists? "/usr/local/bin/sxiv" }
 end
 
 execute "install sxiv" do
@@ -15,4 +16,5 @@ execute "install sxiv" do
 		make install
 	EOH
 	action :nothing
+	creates "/usr/local/bin/sxiv"
 end
